==> GOD/code/01_import-genotypes-join-otlith-transfer-records.Rmd <==
---
title: "01 - (re) Import Genotypes & Join Otolith Transfer Data"
subtitle: "Re-do of `PWS_pink_match_analysis_all_develop.Rmd` with `GCLr`"
author: "Kyle Shedd"
date: "2024-02-07"
output:
  html_notebook:
    theme: united
    toc: yes
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup

Load all necessary packages, as of 2023-08-11 Kyle switched to [GCLr](https://github.com/commfish/GCLr).
```{r setup, message=FALSE, results='hide'}
rm(list = ls(all.names = TRUE))

if(!require("pacman")) install.packages("pacman"); library(pacman)

pacman::p_load(
  tidyverse,
  lubridate,
  scales,
  janitor,
  GCLr
)

knitr::opts_chunk$set(fig.width = 10)

.username = readLines("~/R/usr_pw.txt", n = 1)  # LOKI username
.password = readLines("~/R/usr_pw.txt" , n = 2)[[2]]  # LOKI password
```

# Objective

This is a re-do of `PWS_pink_match_analysis_all_develop.Rmd` with updated heart genotypes, now that all PWS Pink Salmon heart genotyping is finalized as of August 2023.

The purpose of this notebook is to prepare the data for our attempt to re-pair GT-seq genotypes from otolith-derived DNA with GT-seq genotypes from heart-derived DNA for individuals of unknown origin. This notebook will match up the 96 shallow well plate (SWP) locations from `PGOD21.gcl` to the original 48 deep well plate (DWP) to prepare for matching (separate notebook). This will be done by linking 3 things:  

  1) `PGOD21.gcl` otolith genotypes from 96 SWP,
  2) Otolith transfer records documenting the transfer of otoliths from their original 48 DWP (or box of 48 DWP),
  3) AHRP Salmon Biological Fact data containing field data (silly code, stream, year, 48 DWP, etc.)
  
Bottom line, we need to know to the finest scale possible (48 DWP or box) where otoliths in `PGOD21.gcl` came from so we can limit our search space for matching.

# Background

During shipment of otolith samples in February 2020 to the MTA in Juneau for reading, a significant proportion of otoliths migrated between cells within DWPs due to poor containment of the acetate lids that were attached with rubber bands to the DWPs. This incident is known as "the great otolith debacle", aka the "GOD" incident. Since some otoliths moved from their original cells in the DWP, the paired integrity of the otolith-origin information and the rest of the paired data (genotype + field data) was lost for just under 11,000 individuals from the following collections:  

  * PERB15
  * PERB16
  * PERB17
  * PGILMOUR15
  * PGILMOUR16
  * PGILMOUR17
  * PGILMOUR18
  * PPADDY16

In an attempt to rectify the "GOD" incident, we extracted DNA from the otolith tissues, genotyped the otolith-derived DNA at 298 GT-seq loci, and are attempting to re-pair the otolith-heart samples from their genotypes.  

In addition to this GitHub repository, other materials can be found on the V: drive here:
[link](V:\Lab\Loki\Collection raw data sheets\Pink\Otolith Extravaganza\GOD Event 2020_2021)

# Methods

DNA was extracted from left-side or unknown otolith tissues of unknown origin using conventional Machery-Nagel DNA extraction kits. Otoliths were placed in T1 buffer for an overnight soak in clear 96 shallow-well plates (SWP). Preamp at 14 cycles. Final elution volume was 75uL. All otoliths were transferred among plates with a jig. All liquid handling was done by robot.

# Import Data

## Genotypes

### `LocusControl`

Create `LocusControl` using our standard 298 Pink Salmon GT-seq loci.
```{r}
GCLr::load_objects(path = "../objects", pattern = "loci298")

GCLr::create_locuscontrol(locusnames = loci298,
                          username = .username,
                          password = .password)

GCLr::save_objects(objects = "LocusControl", path = "../objects")
```

### Genotypes

Read in `PGOD21` and all heart genotypes for all affected collections (2024-02-05).
```{r}
GOD_affected_heart_sillys <- c(paste0("PERB", 15:17),
                               paste0("PGILMOUR", 15:18),
                               "PPADDY16")

GCLr::loki2r(
  sillyvec = c("PGOD21", GOD_affected_heart_sillys),
  username = .username,
  password = .password, 
  test_type = "GTSNP"
)
```

Save objects
```{r}
if(!dir.exists("../data/genotypes")) {dir.create("../data/genotypes")}
if(!dir.exists("../data/genotypes/raw")) {dir.create("../data/genotypes/raw")}
GCLr::save_sillys(
  sillyvec = c("PGOD21", GOD_affected_heart_sillys),
  path = "../data/genotypes/raw",
  rds = TRUE
)
```

### Reload

Re-load objects if necessary
```{r}
# GCLr::load_objects(path = "../objects", pattern = "loci298")
# GCLr::load_objects(path = "../objects", pattern = "LocusControl")
# GOD_affected_heart_sillys <- c(paste0("PERB", 15:17),
#                                paste0("PGILMOUR", 15:18),
#                                "PPADDY16")
# GCLr::load_sillys(path = "../data/genotypes/raw", rds = TRUE)
```

## Otolith Transfer from MTAL

Read in the otolith transfer data from our trip down to the MTAL in March 2021. This contains the original 48 DWP information and the 96 SWP otolith transfer locations, but no SILLY or collection information. The file was a combination of the DWP transfer file sent by Jodi Neil on 2021-11-17 (`otolith_transfer_96_template_LEFT.xlsx`) and the vial otolith transfer file sent by Jodi Neil on 2021-08-17 (`otolith_transfer_96_template_VIALS_LEFT.xlsx`) and Jodi Estrada's in-house GCL otolith transfers (`JLE_otolith_transfer_96.xlsx` & "PGOD21 data additions" e-mail from 2022-04-27) that Kyle saved as a `.csv` for ease of reading into R.
```{r}
(
  oto_xfer_data_raw <-
    readr::read_csv("../data/otolith_transfer_final.csv")
)
```

Verify that all NTC are in position 96
```{r}
oto_xfer_data_raw %>% 
  dplyr::filter(`96_oto_tray_position` == 96) %>% 
  dplyr::count(`48_DWP_tray_position`)

oto_xfer_data_raw %>% 
  dplyr::filter(`48_DWP_tray_position` == "NTC") %>% 
  dplyr::count(`96_oto_tray_position`)

```

Great, strip out the NTCs
```{r}
oto_xfer_data <- oto_xfer_data_raw %>% 
   dplyr::filter(`96_oto_tray_position` != 96)  # filter out NTC

rm(oto_xfer_data_raw)
```

How many otolith tranfer records? How many otolith genotypes in `PGOD21.gcl`?
```{r}
nrow(oto_xfer_data)
nrow(PGOD21.gcl)
```

**NOTE** we have otolith transfer records for 10,822 otoliths, but only genotypes for 10,791 otoliths in `PGOD21.gcl`

## AHRP Salmon Biological Fact

Read in all of the heart tissue metadata for all PWS Pink Salmon collections from 2013-2020. We'll use this to match up the 48DWP barcodes in otolith transfer records to collection info (silly code, etc.).
```{r}
(
  AHRP_oceanAK <-
    readr::read_csv(file = "../../../OceanAK/AHRP Salmon Biological Data 20221212_162324.csv") %>% 
    janitor::clean_names()
)
```

# Transform Data

We need to join the transfer data with `PGOD21.gcl` so we have the `PGOD21.gcl` otolith genotypes paired with the 48DWP barcode and original silly. This will allow us to compare the `PGOD21.gcl` genotypes to the heart genotypes from the same 48DWP and limit our search image.

## Subset 48DWP & Silly

Make a simple tibble of the 48DWP barcode and silly from `AHRP_oceanAK`
```{r}
(
  AHRP_DWP_data <- AHRP_oceanAK %>%
    dplyr::mutate(
      "48_DWP_tray_barcode" = as.numeric(dna_tray_code),
      sample_date = lubridate::as_date(sample_date)
    ) %>%
    dplyr::rename(
      stream = location_code
    ) %>%
    dplyr::distinct(`48_DWP_tray_barcode`, silly_code, stream, sample_year, sample_date)
)
```

## Make Data Key

Make a data key with `96_oto_tray_barcode` and `96_oto_tray_position` for both the `oto_xfer_data` and `PGOD21.gcl`.

Already, removed NTC wells from `oto_xfer_data`
```{r}
(
  oto_xfer_data <- oto_xfer_data %>%
    # dplyr::filter(!(`48_DWP_tray_barcode` %in% "NTC")) %>%  # already did this above
    tidyr::unite(
      col = "oto_tray_and_well",
      c(`96_oto_tray_barcode`, `96_oto_tray_position`),
      sep = "_",
      remove = FALSE
    )
)

PGOD21.gcl <- PGOD21.gcl %>%
  dplyr::mutate("96_oto_tray_barcode" = as.numeric(DNA_TRAY_CODE)) %>%
  tidyr::unite(
    col = "oto_tray_and_well",
    c(`96_oto_tray_barcode`, `DNA_TRAY_WELL_CODE`),
    sep = "_",
    remove = FALSE
  )
```

# Join Otolith Transfer & 48DWP Data With `PGOD21`

By joining the transfer data we get the original 48DWP, which we can then join to the AHRP OceanAK metadata to get the 48DWP-level sampling data (silly code, stream, year, sample date).
```{r}
PGOD21.gcl <- PGOD21.gcl %>%
  dplyr::left_join(y = oto_xfer_data,
                   by = c("oto_tray_and_well", "96_oto_tray_barcode")) %>%
  dplyr::left_join(y = AHRP_DWP_data, by = "48_DWP_tray_barcode") %>% 
  dplyr::relocate(dplyr::contains("Ogo_"), .after = dplyr::last_col())  # move genotype columns to the end
```

## How Many Fish and How Many 48DWPs?

How many unique 48DPWs have GOD otolith genotypes?
```{r}
dplyr::n_distinct(PGOD21.gcl$`48_DWP_tray_barcode`, na.rm = TRUE)

# Count of 48DWP per silly
PGOD21.gcl %>% 
  dplyr::distinct(silly_code, `48_DWP_tray_barcode`) %>% 
  dplyr::count(silly_code)

# Count of fish per silly
PGOD21.gcl %>% 
  dplyr::count(silly_code) %>% 
  janitor::adorn_totals()
```

There are **439 48DWPs** involved in GOD!

What about **160 otoliths** that did not assign to a 48DWP or silly? Are they all *box* otoliths that were just kicking around in one of the shipping boxes and were later put into vials at the MTAL? Or do we potentially have some data entry errors?
```{r}
PGOD21.gcl %>% 
  dplyr::filter(is.na(silly_code)) %>% 
  dplyr::mutate(is_box = !is.na(box_number)) %>% 
  dplyr::count(is_box)
```

The transfer records for the **159 box otoliths** showing which 48DWPs were in which boxes are here `Oto Boxes that shipped to JNU.xlsx`.

But the remaining issue here is that, there is **1 otolith** that does not match up to a 48DWP or `box_number`, WTF?!?!?!
```{r}
PGOD21.gcl %>% 
  dplyr::filter(is.na(`48_DWP_tray_barcode`),
                is.na(`box_number`)) %>% 
  dplyr::select(-dplyr::contains("Ogo_")) %>% 
  dplyr::count(DNA_TRAY_CODE)
```

It looks like this 1 otolith is on tray `0000050960`. No clue what these are, still need to resolve...

Are these otolith trays present in the transfer data?
```{r}
PGOD21.gcl %>% 
  dplyr::anti_join(oto_xfer_data, by = "oto_tray_and_well") %>% 
  dplyr::count(`96_oto_tray_barcode`, PLATE_ID)  # in PGOD21.gcl, not in transfer records
```

What about otoliths in the transfer records, but not genotyped in `PGOD21.gcl`?
```{r}
oto_xfer_data %>% 
  dplyr::anti_join(PGOD21.gcl, by = "oto_tray_and_well") %>% 
  dplyr::count(`96_oto_tray_barcode`)  # in transfer records, but not in PGOD21.gcl
```

This partial tray is in the otolith transfer records, but shows it was only filled until position 21. However, we have genotypes for positions 22 (no data) and 25-33 (Jodi Estrada's transfers).

How many otolith genotypes do we have in the partial?
```{r}
PGOD21.gcl %>% 
  dplyr::filter(DNA_TRAY_CODE == "0000050960")
```

Jodi Estrada also transferred some right-side otoliths into 50943, did those get genotyped?
```{r}
PGOD21.gcl %>% 
  dplyr::filter(DNA_TRAY_CODE == "0000050943")
```

Nope, it does not look like we genotyped those right side otoliths.

# Oto Transfer Comments

We'll need to review these in the course of re-pairing otoliths with hearts.
```{r}
oto_xfer_data %>% 
  dplyr::count(comment)
```

# Save `PGOD21` Data Pairing

```{r}
PGOD21.gcl %>% 
  dplyr::select(-dplyr::contains("Ogo_")) %>% 
  dplyr::glimpse()
```

Strip to the bare essentials and save.
```{r}
(
  PGOD21_48DWP_paired <- PGOD21.gcl %>%
    dplyr::select(
      SillySource,
      oto_tray_and_well,
      `96_oto_tray_barcode`:sample_date
    )
)

readr::write_csv(x = PGOD21_48DWP_paired, file = "../output/PGOD21_48DWP_paired.csv")
```

# Final Stats 2024-02-07

10,791 otoliths with genotypes.

```{r}
PGOD21_48DWP_paired %>%
  dplyr::mutate(
    match_type = dplyr::case_when(
      !is.na(`48_DWP_tray_barcode`) & !is.na(`48_DWP_tray_position`) ~ "DWP+Position",
      !is.na(`48_DWP_tray_barcode`) & is.na(`48_DWP_tray_position`) ~ "DWP_Only",
      is.na(`48_DWP_tray_barcode`) & !is.na(box_number) ~ "Box",
      is.na(`48_DWP_tray_barcode`) & is.na(box_number) ~ "Nothin",
      TRUE ~ "WTF"
    )
  ) %>% 
  dplyr::count(match_type)
```

Got it down to 1 unaccounted for otolith on tray `50960`, position 22.

# Save `PGOD21` otolith genotypes

Save objects
```{r}
if(!dir.exists("../data/genotypes/raw_join_48DWP")) {dir.create("../data/genotypes/raw_join_48DWP")}
GCLr::save_sillys(
  sillyvec = "PGOD21",
  path = "../data/genotypes/raw_join_48DWP",
  rds = TRUE
)
```

# Subset and Save Heart Genotypes

Filter DWPs for only affected DWPs
```{r}
affected_DWPs <- unique(PGOD21.gcl$`48_DWP_tray_barcode`)

AHRP_DWP_data %>% 
  dplyr::filter(`48_DWP_tray_barcode` %in% affected_DWPs)
```

There are **439** GOD-affected DWPs according to the otolith transfer records.

Pool all heart genotypes, then filter for otolith DWPs.
```{r}
GCLr::pool_collections(collections = GOD_affected_heart_sillys, newname = "heart_genotypes")
```

Filter heart genotypes for affected DWPs. **NOTE** this assumes that any *box* otoliths would have come from one of the affected DWPs.
```{r}
heart_genotypes.gcl <- heart_genotypes.gcl %>%
  dplyr::mutate("48_DWP_tray_barcode" = as.numeric(DNA_TRAY_CODE)) %>% 
  dplyr::relocate(dplyr::contains("Ogo_"), .after = dplyr::last_col())  # move genotype columns to the end

heart_genotypes_IDs_to_remove <- heart_genotypes.gcl %>%
  dplyr::filter(!`48_DWP_tray_barcode` %in% affected_DWPs) %>% 
  dplyr::pull(FK_FISH_ID)

GCLr::remove_ids(silly = "heart_genotypes", IDs = heart_genotypes_IDs_to_remove)
```

How many heart genotypes are left?
```{r}
nrow(heart_genotypes.gcl)
```

Save
```{r}
GCLr::save_sillys(
  sillyvec = "heart_genotypes",
  path = "../data/genotypes/raw_join_48DWP",
  rds = TRUE
)
```

Make sure that all GOD affected DWPs are included, should have 439 DWPs.
```{r}
heart_genotypes.gcl %>% 
  dplyr::distinct(`48_DWP_tray_barcode`, DNA_TRAY_CODE)
```

Uh oh, it looks like we only have heart genotypes for 438 out of the 439 GOD-affected DWPs.

## Missing Heart Genotypes?

Huh, I appear to only have 438 out of 439 DWPs, what gives? Better cross reference the outersect with all AHRP DWPs, maybe some are not GOD-related, but other otolith incidents.
```{r}
missing_heart_DWPs <- setdiff(affected_DWPs, unique(heart_genotypes.gcl$`48_DWP_tray_barcode`))

AHRP_DWP_data %>% 
  dplyr::filter(`48_DWP_tray_barcode` %in% missing_heart_DWPs)
```

It is very weird that 14360 (PPADDY16) has not been genotyped, as that entire SILLY has been genotyped, none remaining. What is the deal here? Let's check out the original data for this DWP.
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000014360")
```

Ah ha! this DWP only had 1 fish sampled, with unknown sex. Since we only extracted DNA from samples with known sex, it makes sense that this DWP was not selected for extraction/genotyping. 

How many GOD otoliths are linked to this DWP?
```{r}
PGOD21.gcl %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 14360)
```

Well shoot, this is unexpected and unwelcome. We have 11 otoliths matching back to this 1 DWP with only 1 samples (which we didn't genotype the heart...). I went back to the raw otolith transfer file sent by Jodi Neil on 2021-08-17 (`otolith_transfer_96_template_VIALS_LEFT.xlsx`), and confirmed that there are indeed 11 otoliths that appear to be from this DWP. Not sure what is going on here.

# Things to Consider for Matching

Before we finalize otolith-heart "matches", there are a few modifications that I'd like to make:  

  * we need a genotyping success threshold for otoliths, it doesn't have to be very high (likely ~5-10%), but we need something to weed out matches where we only have 5 loci for an otolith  
  * we need to implement a heterozygosity filter for hearts, just like we do for parentage analysis, otherwise we'll end up with some very heterozygous heart genotypes that will match *very* well to our highly contaminated (read heterozygous) otoliths  
  * will want to check otolith transfer comments prior to finalizing matches  
  * will want to review `raw_working_oto_sheets/working_oto_sheet_combined.xlsm` to see DWP map prior to finalizing  

# Next Step

Quality assurance for both otolith and heart genotypes!  

  * Read in genotypes we've saved in `data/genotypes/raw_join_48DWP`
  * Standard QA for hearts
  * Only missing loci and duplicate check for otoliths

End
