==> GOD/code/05_identify-wells-in-GOD.Rmd <==
---
title: "05 - Identify Wells In GOD"
subtitle: "Which DWP Wells Have Unpaired Otolith Data?"
author: "Kyle Shedd"
date: "Started: 2025-03-19, last opened: `r Sys.Date()`"
output:
  html_notebook:
    theme: united
    toc: yes
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup

```{r setup, message=FALSE, results='hide'}
rm(list = ls(all.names = TRUE))

if(!require("pacman")) install.packages("pacman"); library(pacman)

pacman::p_load(
  tidyverse,
  scales,
  janitor,
  GCLr
)

knitr::opts_chunk$set(fig.width = 10)

.username = readLines("~/usr_pw.txt", n = 1)  # LOKI username
.password = readLines("~/usr_pw.txt" , n = 2)[[2]]  # LOKI password
```

# Objective

The purpose of this notebook is to identify which DWP wells were part of GOD and have untrustworthy otolith pairing with hearts. We already have the DWP well data for the GOD otolith transfers 48 DWP --> 96 SWP (i.e., where they were removed from at MTAL for genotyping/otolith reading in `GOD/output/PGOD21_48DWP_paired_genotype_matches.csv`), but we do not have a tidy, data friendly version of where they started. We will extract this info from the DWP maps in `GOD/data/raw_working_oto_sheets.xlsm`, which denote GOD wells with red font.

The next steps after this will be to *investigate GOD otoliths with reads but no heart match* to make additional assignments based on logic if all remaining otolith reads are the same, then *verify heart matches are GOD fish (i.e., don't have existing otolith reads)*, and ultimately *submit finalized PGOD21 matches to Eric Lardizabal*. 

# Background

During shipment of otolith samples in February 2020 to the MTA in Juneau for reading, a significant proportion of otoliths migrated between cells within DWPs due to poor containment of the acetate lids that were attached with rubber bands to the DWPs. This incident is known as "the great otolith debacle", aka the "GOD" incident. Since some otoliths moved from their original cells in the DWP, the paired integrity of the otolith-origin information and the rest of the paired data (genotype + field data) was lost for just under 11,000 individuals from the following collections:  

  * PERB15
  * PERB16
  * PERB17
  * PGILMOUR15
  * PGILMOUR16
  * PGILMOUR17
  * PGILMOUR18
  * PPADDY16

In an attempt to rectify the "GOD" incident, we determined which otoliths were "suspect" and had likely migrated between wells, transferred left and unknown otoliths to 96 shallow well plates (SWP), extracted DNA from the otolith tissues, genotyped the otolith-derived DNA at 298 GT-seq loci, and are attempting to re-pair the otolith-heart samples from their genotypes. 

In addition to this GitHub repository, other materials can be found on the V: drive here:
[link](V:\Lab\Loki\Collection raw data sheets\Pink\Otolith Extravaganza\GOD Event 2020_2021)

# Methods

Here's a running list of methods and brief summary of results to help me remember how we got here since there are large gaps in time between working on this project.

## Otolith Tranfers

We determined which otoliths were "suspect" and had likely migrated between wells based on comparing the DWP otolith inventory at the MTAL (destination, post-GOD) to the the inventory at the GCL (origin, pre-GOD). Full documentation of the *rules* are located here: `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`, tab "decisions". Briefly, any well that had an inventory discrepancy was "bad" and any adjacent well was "suspect" and physically labelled along the left edge of the well.

Only **left or unknown** otoliths from "suspect" wells were transferred to 96 SWPs fro reading (i.e., standard otolith trays) with 1 otolith per well. Full documentation of the *rules* are located in the Excel files here: `data/raw_otolith_transfer_96`, see tab "oto_transfer_business_rules". Right otoliths were also transfered, but not read. After completing the transfers at the MTAL, SWPs were sent to the GCL for DNA extraction and genotyping.

## Laboratory

DNA was extracted from `PGOD21` left-side or unknown otolith tissues of unknown origin using conventional Machery-Nagel DNA extraction kits. Otoliths were placed in T1 buffer for an overnight soak in clear 96 SWPs. Preamp at 14 cycles. Final elution volume was 75uL. All otoliths were transferred among plates with a jig. All liquid handling was done by robot. Samples were genotyped using GT-seq in lab projects P062, P065, P066, P067, P068, P069, P070, P071, P072, P073. QC samples were included in P067, P068, and P074.

Projects P070-072 included an additional step to dry down extracted DNA to increase concentrations and genotyping success, however, this did not occur in P062-P069 and P073. There is a slight difference in average genotyping rate among these projects (75% vs. 68%), but insufficient to warrant re-running P062-P069 and P073. See e-mails from 2024-11-27 for more info.

## Analysis Workflow

### Data Prep

All data prep was done in the R Notebook `01_import-genotypes-join-otlith-transfer-records`. The purpose of that R Notebook was to come up with a definitive link between the `PGOD21.gcl` otoliths and their 96 SWP locations and the *original* 48 DWP or otolith box location. This was accomplished by linking these 3 things:  

  1) `PGOD21.gcl` otolith genotypes from 96 SWP,
  2) Otolith transfer records documenting the transfer of otoliths from their original 48 DWP (or box of 48 DWP),
  3) AHRP Salmon Biological Fact data containing field data (silly code, stream, year, 48 DWP, etc.)
  
Bottom line, we need to know to the finest scale possible (48 DWP or box) where otoliths in `PGOD21.gcl` came from so we can limit our search space for matching.

How many otoliths with genotypes in `PGOD21.gcl`? By collection?

Final stats for `PGOD21.gcl` from `01_import-genotypes-join-otlith-transfer-records`:

 silly_code        n  
     PERB15      478  
     PERB16    1,527  
     PERB17      525  
 PGILMOUR15    1,587  
 PGILMOUR16    1,555  
 PGILMOUR17    2,600  
 PGILMOUR18      856  
   PPADDY16    1,503  
       Box       159  
Unaccounted        1  
      Total   10,791  

match_type         n  
          Box    159  
 DWP+Position  9,623  (updated in 04_)
     DWP_Only  1,008  (updated in 04_)
       Nothin      1  
        Total 10,791  

**439 unique 48 DWPs**

### Quality Assurance (QA)

All QA was done in the R Notebook `02_genotype-QA-hearts-otoliths`. The purpose of that R Notebook was to perform genotyping quality assurance (QA) on the otoliths in PGOD21.gcl and GOD-affected hearts in Erb 2015-2017, Gilmour 2015-2018, and Paddy 2016.

Hearts undergo standard AHRP PWS Pink Salmon QA:  

  1) missing genotypes (<80% loci)
  2) duplicate check (>95% loci, remove both)
  3) contamination (heterozygosity outliers, +/- 3.5 modified z-score, separate for even/odd lineage)

Otoliths undergo a revised QA:  

  1) missing genotypes (<15% loci, arbitrary cutoff)
  2) duplicate check (>99% loci[contamination + low genotyping success], remove 1 [keep sample w/ most location info and genotype data])
  3) contamination (<20 non-heterozygous loci, all together not separate by lineage)

We started with 19,898 hearts from 8 collections. We removed 1,180 (5.9%) due to missing, 60 (0.6%) due to duplicates (both removed), and 592 (3.0%) due to contamination. Final set of hearts is 18,066 (90.8%).

We started with 10,791 otoliths in PGOD21. We removed 522 (4.8%) due to missing, 87 (0.8%) due to duplicates (only 1 removed), and 785 (7.3%) due to contamination. Final set of otoliths is 9,397 (87.1%).

### Duplicate Check

After QA, we filtered down to a common set of 435 DWPs with 9,258 otoliths (85.8%) and 17,943 hearts (90.2%) and calculated the duplicate rate (% of loci sharing genotypes) for all otoliths and hearts within a DPW for a total of 394,316 total pairwise comparisons.

### Matching Prep

Techniques for matching otoliths back to hearts were refined over time, however, the most definitive work is in this R Notebook `Known sample tests_Kyle.Rmd`, also titled `Matching tissues to otoliths with known samples`, which revisited T028 data to see if `CKMRsim` was better for matching tissues than a standard duplicate check. Andy Barclay had the idea to try `CKMRsim`, and while initial results were promising, they ultimately fared no better than a duplicate check.

Bottom line, this R notebook refined the methods used for otolith-heart matching. Using cutoffs for duplicate rate >= 0.8 and diff (best match - 2nd best match) >= 0.25 appears to maximize the number of correct matches while minimizing the number of incorrect matches.

### Previous Matching

Previous matching was completed by Kristen in `old/PWS_pink_match_analysis_all.rmd` with results saved in `../dup_check_results_P066-P067_pass_with_wells.csv`. This is the latest modified file named "dup_check_results*", however, the name of the output file is a bit confusing given that it should have included more than just those two projects. 

However, there are some old R Notebooks that are also titled `old/PWS_pink_match_analysis_P062-P065.Rmd` and `old/PWS_pink_match_analysis_P066-P067.Rmd`. Need to dig into these to see where their output is.

### Current Matching

We took the within DWP duplicate checks for 9,258 post-QA otoliths to determine the highest `duplicate_rate`  and `diff` between duplicate rates for the 1st best and 2nd best match. Based on results from T028 in `Known sample tests_Kyle.Rmd`, Kristen and I decided to use threshold cutoffs of `duplicate_rate >= 0.8` and `diff >= 0.25` for assigning matches, as this minimized false positives. There is a .pdf in `output` with visualizations of how otoliths shifted within 48 DWPs, which can be compared to the original DWP otolith inventories `raw_working_oto_sheets`. No otoliths above this threshold matched > 1 heart (i.e., no ties). There were 306 hearts that were matched to 2 or 3 otoliths above threshold cutoffs. Of those, we kept 226 hearts that matched to otoliths with the same otolith reads (identical `mark_present` and `mark_id`) and the remaining matches were dropped. In total, **5,421** out of 9,258 post-QA otoliths were assigned to a heart based on genotype data (50.2%). Of those, the vast majority (3,947; 72.8%) had not moved (otolith and heart in same well).

# Import Data

## DWP Maps

Get the DWP maps data so I can relate the Excel cells to DWP barcode and position.

**NOTE** `../data/raw_working_oto_sheets/DWP_maps_clean.xlsm` is a copy from `../data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`.
```{r}
(
  DWP_maps <- readxl::read_excel(
    path = "../data/raw_working_oto_sheets/DWP_maps_clean.xlsm",
    sheet = "DWP maps",
    range = "A1:G8139",  # updated after fabricating DWP maps for Jodi's in-house DWP transfers
    col_names = c("row_DWP", "col_1", "col_2", "col_3", "col_4", "col_5", "col_6")
  )
)
```

## GOD DWP Wells

Get the Excel cells with red font that denote GOD DWP wells. **NOTE** I had ChatGPT write a VBA macro for me to make a list of excel cells with red font. Additional details and the VBA code are in the "README" worksheet in `../data/raw_working_oto_sheets/DWP_maps_clean.xlsm`

VBA macro last run 2025-04-02!
```{r}
(
  red_cells <- readxl::read_excel(
    path = "../data/raw_working_oto_sheets/DWP_maps_clean.xlsm",
    sheet = "RedFontCells",
    col_names = "cell", 
    skip = 1  # skip header
  )
)
```

Total of 11,400 (original was 10,898) GOD DWP wells as of *2025-04-02*. This is close, but not the same as the 10,791 PGOD21 otoliths.

## AHRP Salmon Biological Fact

Need to read in the latest and greatest otolith data from the warehouse as of 2025-03-03.
```{r}
(
  AHRP_oceanAK <-
    readr::read_csv(file = "../../../OceanAK/AHRP Salmon Biological Data 20250303_092938.052746.csv") %>% 
    janitor::clean_names() %>% 
    tidyr::unite(col = "heart_fish_id", silly_code:fish_id, sep = "_", remove = FALSE) %>%  # add heart SillySource
    dplyr::mutate(sample_date = lubridate::as_date(sample_date))  # fix sample_date
)
```

## Stream Specimens

Read in the raw contractor data, it notes whether any otoliths were missing in the field.
```{r}
(
  stream_specimens <- readr::read_csv(file = "../../../Stream Specimens/StreamSpecimens_PWS_2013-2020_CSVReport.csv", 
                                      show_col_types = FALSE) %>%
    janitor::clean_names()
)
```

## Otolith Transfer & Matching Data

Read in the otolith transfer data that matches up 96 well otolith tray barcodes and wells to the original 48 DWP barcodes and wells. This was generated in `01_import-genotypes-join-otlith-transfer-records` with duplicate rate data used to assign matches in `04_matching-hearts-otoliths`.

Eric Lardizabal ultimately needs to know for each otolith:  

  1) `Target Barcode` = `48_DWP_tray_barcode` (10-digit heart 48DWP barcode)  
  2) `Target Well Position` = `heart_position` (heart 48DWP well position, i.e., A1, A2, etc.)

The two other GOD columns in the *LOKI* tissue table are:  

  * `Determination Collection` = "PGOD21", only used in the target collection (i.e., in PPADDY16 the Determination Collection = PGOD21)  
  * `Target Container Type` = type of container the otolith is going into, should be 96 well SWP

Recall that only left and unknown otoliths were transferred. No right-side otoliths were transferred.
```{r}
(PGOD21_48DWP_paired_genotype_matches <- readr::read_csv(file = "../output/PGOD21_48DWP_paired_genotype_matches.csv") %>% 
   dplyr::mutate(oto_side = stringr::str_to_lower(oto_side)))
```

Total of 10,791 otoliths in `PGOD21.gcl`, 5,421 of which were assigned to a heart based on genetic data.

# Translate Excel Cell to DWP Well

## Columns

First step, split Excel cells into row and column, then get DWP column based on excel column.
```{r}
excel_col_to_num <- function(col) {
  sapply(col, function(x) {
    sum((match(stringr::str_split(x, "")[[1]], LETTERS)) * 26^(seq_along(stringr::str_split(x, "")[[1]]) - 1))  # thanks ChatGPT
  })
}

(
  red_cells_mod <- red_cells %>%
    dplyr::mutate(
      col_excel = stringr::str_extract(cell, "^[A-Z]+"),  # Extract letters (column)
      row_excel = as.numeric(stringr::str_extract(cell, "\\d+$")),  # Extract numbers (row)
      col_DWP = excel_col_to_num(col_excel) - 1  # B = 1, C = 2, D = 3, E = 4, F = 5, G = 6
    )
)
```

## Rows

Okay, now for the slightly harder part, translating the Excel row data into DWP row data.
```{r}
# Base R

row_DWP_vec <- DWP_maps %>% dplyr::pull(row_DWP)  # just get the vector

n_DWP <- (length(row_DWP_vec) + 1) / 11  # 11 rows per DWP: barcode, A:H, sum oto, NA

DWP_barcode_i <- seq(from = 1, by = 11, length.out = n_DWP)

DWPs <- row_DWP_vec[DWP_barcode_i]
 
# red_cells_mod$row_DWP <- row_DWP_vec[red_cells_mod$row_excel]

# tidyverse
(
  red_cells_mod <- red_cells_mod %>%
    dplyr::left_join(
      y = DWP_maps %>%
        dplyr::mutate(row_excel = dplyr::row_number()) %>% 
        dplyr::select(row_excel, row_DWP),
      by = "row_excel"
    )
)
```

Confirm all equal base R vs. tidyverse.
```{r}
all.equal(red_cells_mod$row_DWP, row_DWP_vec[red_cells_mod$row_excel])
```

## DWP

And finally, associating the DWP wells with a DWP barcode.
```{r}
# Base R
(
  red_cells_mod <- red_cells_mod %>%
    dplyr::mutate(`48_DWP_tray_barcode` = DWPs[(row_excel %/% 11) + 1]) %>%  # integer division - get DWP barcode all rows below, repeat by 11
    tidyr::unite(col = "heart_position", row_DWP:col_DWP, sep = "")
)
```

# Filter For Red Cells With Number

I noticed that some Excel cells with red font were blank, no otolith data. This is likely due to copy/paste leaving some ghost formatting (red font cells, no data). Need to purge these so we get an accurate count!

Create a tidy version of `DWP_maps`, one row per cell.
```{r}
(
  DWP_maps_tidy <- DWP_maps %>%
    dplyr::mutate(
      dplyr::across(dplyr::everything(), as.character),
      row_excel = dplyr::row_number()
    ) %>%
    dplyr::rename(
      A = row_DWP,
      B = col_1,
      C = col_2,
      D = col_3,
      E = col_4,
      F = col_5,
      G = col_6
    ) %>%
    tidyr::pivot_longer(
      cols = A:G,
      names_to = "col_excel",
      values_to = "value"
    ) %>%
    tidyr::unite(col = "cell", col_excel:row_excel, sep = "", remove = TRUE)
)
```

Filter for non-NA, numeric values.
```{r}
(
  red_cells_mod <- red_cells_mod %>%
    dplyr::select(-col_excel, -row_excel) %>% 
    dplyr::left_join(y = DWP_maps_tidy, by = "cell") %>%
    dplyr::mutate(value = as.numeric(value)) %>%  # force to numeric
    dplyr::filter(!is.na(value))
)
```

A total of **11,260** GOD wells, as of 2025-04-02.

Verify nothing weird.
```{r}
red_cells_mod %>% 
  dplyr::count(value) %>% 
  dplyr::arrange(value)
```

Wow, some DWP wells had a LOT of otoliths in them. Yikes!

# Verify PGOD21 Assignments Are GOD Wells

First we need to make sure that we haven't assigned any PGOD21 otoliths to non-GOD DWP wells. This will be a good double check to make sure we have **ALL** the GOD wells.
```{r}
# wells
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = red_cells_mod %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position))

# count of wells by DWP
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = red_cells_mod %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
  dplyr::count(`48_DWP_tray_barcode`) %>% 
  dplyr::arrange(dplyr::desc(n))
```

Shit, as of 2025-03-27 there are 290 assignments from 76 DWPs to non-red cells.
2025-03-27 down to 170 assignments from 65 DWPs after fixing red font issue.
2025-03-28 down to 77 assignments from 54 DWPs after adding Jodi's 7 in-house transfers.

How many of these are DWP+Position vs. vial?
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = red_cells_mod %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
  dplyr::count(otolith_info)
```

Interesting, 71 are DWP+Position and only 6 are vial otoliths.

This means that we have PGOD21 assignments to non-red cells. Need to go back into `data/raw_working_oto_sheets/DWP_maps_clean.xlsm` and add red formatting and re-run the VBA macro.

Use these code chunks to investigate individual DWPs.

Assignments
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = red_cells_mod %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 3088) %>% 
  dplyr::arrange(heart_position)
```

Otolith reads
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000022683") %>% 
  dplyr::select(dna_tray_code, dna_tray_well_code, otolith_mark_present, otolith_mark_id, otolith_mark_status_code)
```

Some had the wrong red font, so the VBA didn't work, others did not have red font at all. Also noticed some DWPs with GOD issues, but 0 red cells and 0 otoliths transferred...7 of the 8 DWPs that Jodi Estrada transferred in-house never got DWP maps, so I made them up based on the transfer records and existing otolith reads. Those 7 DWPs were:  

    3088
    2413
    3255
    2381
    22262
    3079
    2372

Going to leave at 77 assignments from 54 DWPs until we finish investigating GOD samples with existing otolith reads. From there we can revisit this.

# Verify Otolith Transfers In GOD Wells

One way to check whether any entire DWPs are missing from the GOD wells is to cross-reference the DWPs in the otolith transfer records.

Any DWPs that were transferred, but not identified with red cells in "DWP maps"?
```{r}
dplyr::anti_join(
  x = PGOD21_48DWP_paired_genotype_matches %>%
    dplyr::distinct(`48_DWP_tray_barcode`) %>% 
    dplyr::filter(!is.na(`48_DWP_tray_barcode`)),
  y = red_cells_mod %>%
    dplyr::distinct(`48_DWP_tray_barcode`) %>%
    dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
  by = "48_DWP_tray_barcode"
) %>% 
  dplyr::arrange(`48_DWP_tray_barcode`)
```

## Investigate

### 2839

Only 1 vial otolith transferred, no red cells in "DWP maps", discordant GOD assignment and existing otolith read. Decided to mark red cells and will ask for otolith data to be deleted.

### 2924

Only 1 vial otolith transferred, no red cells in "DWP maps". No GOD assignment. Decided to keep existing reads.

### 14360

11 vial otoliths transferred...but only 1 sample actually collected in the field. Probably an issue with the otolith transfer data.

### 14597

Only 1 vial otolith transferred, no red cells in "DWP maps". GOD assignment matches existing read. Decided to keep existing reads.

### 16127

Only 3 vial otoliths transferred, no red cells in "DWP maps". 1 GOD assignment, C5, 35, does not match existing read. Decided to mark C5 as red and will ask for otolith data to be deleted. Keep existing reads except C5.

### 22517

Only 1 vial otolith transferred, no red cells in "DWP maps". No GOD assignment. Keep existing reads.

### 22675

Only 1 vial otolith transferred, no red cells in "DWP maps". 1 GOD assignment, A4, 25, wild. All neighboring wells are wild. No existing read for A4. Decided to mark A4 red.

### 22696

Only 1 vial otolith transferred, no red cells in "DWP maps". No GOD assignment, didn't pass QA. Keep existing reads, do not mark any red cells.

### 22704

Only 1 vial otolith transferred, no red cells in "DWP maps". No GOD assignment, didn't meet thresholds. Keep existing reads, do not mark any red cells.

Otolith transfers.
```{r}
# PGOD21_48DWP_paired_genotype_matches %>% 
#   dplyr::filter(`48_DWP_tray_barcode` == 22704) %>% 
#   dplyr::arrange(heart_position)
```

Otolith transfer and assignments.
```{r}
# PGOD21_48DWP_paired_genotype_matches %>% 
#   dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
#   dplyr::anti_join(y = red_cells_mod %>% 
#                      dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
#                    by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
#   dplyr::filter(`48_DWP_tray_barcode` == 22704) %>% 
#   dplyr::arrange(heart_position)
```

Otolith reads
```{r}
# AHRP_oceanAK %>% 
#   dplyr::filter(dna_tray_code == "0000022704") %>% 
#   dplyr::select(dna_tray_code, dna_tray_well_code, otolith_mark_present, otolith_mark_id, otolith_mark_status_code)
```

Contractor data, check for missing otoliths in the field data.
```{r}
# stream_specimens %>% 
#   dplyr::filter(sample_tray_id == "0000022696",
#                 missing_otoliths > 0
#                 ) %>% 
#   dplyr::arrange(sample_cell) %>% 
#   dplyr::select(sample_tray_id, sample_cell, missing_otoliths)
```

# Compare GOD Wells to Salmon Biological Fact

Need to do some QA/QC to:  

  1) verify that all of the GOD_48DWP_wells are associated with samples (i.e., otoliths migrated into unsampled wells on partial DWPs), and  
  2) make sure that there are no otolith reads for these GOD DWP wells  

## Join GOD Wells with Otolith Reads

Need to convert from `heart_position` A1:H6 to `dna_tray_well_code` 1:48.
```{r}
pos_to_code <- dplyr::tibble(pos = paste0(rep(LETTERS[1:8], times = 6), rep(1:6, each = 8)), 
                             dna_tray_well_code = as.numeric(1:48))

# join to get dna_tray_well_code, then join to get otolith data from AHRP_oceanak
(
  GOD_48DWP_wells_oceanak <- red_cells_mod %>% 
    dplyr::select(`48_DWP_tray_barcode`, heart_position) %>%
    dplyr::mutate(
      dna_tray_code = stringr::str_pad(string = `48_DWP_tray_barcode`, side = "left", width = 10, pad = "0")
    ) %>%  # need 10 digit barcode, including leading 0's
    dplyr::left_join(y = pos_to_code, by = dplyr::join_by(heart_position == pos)) %>%
    dplyr::left_join(
      y = AHRP_oceanAK,
      by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
    )
)

GOD_48DWP_wells_oceanak %>% 
  dplyr::count(otolith_mark_present, otolith_mark_status_code)
```

Houston, we have a problem. There are *at least 194 GOD samples that have an existing otolith read*. None of these should have otolith reads...

How many GOD samples per collection?
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::count(silly_code)
```

46 samples don't have a silly code?!?!?

## Investigate GOD Samples Without SILLY

WTF are these 46 samples without a `silly_code`? Investigate by DWP barcode.
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(is.na(silly_code)) %>% 
  dplyr::count(`48_DWP_tray_barcode`) %>% 
  dplyr::arrange(dplyr::desc(n))
```

### NA

Was due to DWP 2192 not being recorded correctly. Code and output below are from *before fix*.

What are these NA's?
```{r}
# red_cells_mod %>% 
#   dplyr::filter(is.na(`48_DWP_tray_barcode`))
```

Oh shoot! There is a DWP that was unlabeled (likely forgot to scan the barcode)

See if we can determine from elimination by looking at the otolith transfer records `PGOD21_48DWP_paired_genotype_matches`.

Are there any DWPs in the otolith transfer records that fit this description?

How about any DWPs in the otolith transfer records that do not exist in `DWP_maps_clean.xlsm`?
```{r}
# dplyr::anti_join(
#   x = PGOD21_48DWP_paired_genotype_matches %>%
#     dplyr::distinct(`48_DWP_tray_barcode`),
#   y = red_cells_mod %>%
#     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)) %>% 
#     dplyr::distinct(`48_DWP_tray_barcode`),
#   by = "48_DWP_tray_barcode"
# ) %>% 
#   dplyr::arrange(`48_DWP_tray_barcode`)
```

Well that is not expected...at least it narrows it down to 29 potential DWPs. Based on `DWP_maps_clean.xlsm`, I wonder if it is actually 2192?

Check AHRP Salmon Biological Fact to see if we have otolith reads from 2192.
```{r}
# AHRP_oceanAK %>% 
#   dplyr::filter(dna_tray_code == "0000002192") %>% 
#   dplyr::select(heart_fish_id, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code)
```

The pattern of reads matches, with the exception of having a read for `dna_tray_well_code` = 20 (D3). There should **not** be any reads for that well since it had 0 otoliths (if this `NA` DWP is truly 2192).

Tried looking for pictures of unpacking the DWPs <V:\Lab\Loki\Collection raw data sheets\Pink\Otolith Extravaganza\GOD Event 2020_2021\Pictures>, but could not find this DWP. According to `working_oto_sheet_combined.xlsm`, this DWP would have been unpacked by Kristen on 2021-03-04, but I can't find a picture.

Went back to `working_oto_sheet_combined.xlsm`, checked the `oto unpacking data` worksheet and saw that Kristen went from 2193 to 2192 to 2202. Given that, I feel comfortable upating this to 2192. Updated `working_oto_sheet_combined.xlsm` and `DWP_maps_clean.xlsm` to assign 2192.

### 28756

What is going on with 28756?

According to the `DWP maps`, this DWP has 32 GOD wells...but somehow 10 of them do not have a `silly_code`?
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 28756) %>% 
  dplyr::arrange(dna_tray_well_code)  # sort by what's missing data
```

Wells 31-40 do not exist.

Check the "raw" data from Salmon Biological Warehouse.
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000028756")
```

Only has data for wells 1-30.

As a double check, look at the "raw" contractor data.
```{r}
stream_specimens %>% 
  dplyr::filter(sample_tray_id == "0000028756") %>% 
  dplyr::arrange(sample_cell)
```

Okay, samples 31-40 do not exist.

I conclude that 28756 was a partially filled DWP with samples in wells 1-30, however, during shipment to MTAL, otoliths from wells 9-30 migrated into wells 9-40. DWP column 1 (wells 1-8) didn't have any otoliths move, and DWP column 6 (wells 41-48) was well sealed with rubber bands preventing any otoliths from entering those wells.

Need to **purge** these wells from `GOD_48DWP_wells_oceanak`, since we want to know which samples have unpaired heart/otolith data.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!(`48_DWP_tray_barcode` == 28756 & is.na(silly_code)))
```

### Are DWPs With NA Silly Also Partial DWPs?

Check and see if these other GOD samples without a silly are also partial plates with otoliths migrating to unsampled wells.
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::group_by(`48_DWP_tray_barcode`) %>% 
  dplyr::filter(any(is.na(silly_code))) %>% 
  dplyr::arrange(`48_DWP_tray_barcode`, dna_tray_well_code)
```

Are the NA silly wells great than the max well sampled?
```{r}
# get vector of DWP's with NA silly_code
DWP_with_NA_silly <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(is.na(silly_code)) %>% 
  dplyr::distinct(`48_DWP_tray_barcode`) %>% 
  dplyr::pull()

# get max well sampled from contractor field data
DWP_with_NA_silly_max_well <- stream_specimens %>% 
  dplyr::filter(sample_tray_id %in% stringr::str_pad(string = DWP_with_NA_silly, width = 10, side = "left", pad = "0")) %>% 
  dplyr::slice_max(sample_cell, by = sample_tray_id, n = 1) %>% 
  dplyr::select(sample_tray_id, sample_cell)

# join in max well sampled to see if GOD wells are > max well sampled
GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(is.na(silly_code)) %>% 
  dplyr::left_join(y = DWP_with_NA_silly_max_well, by = dplyr::join_by(dna_tray_code == sample_tray_id)) %>% 
  dplyr::mutate(unsampled = dna_tray_well_code > sample_cell) %>% 
  dplyr::count(unsampled)
```

Okay, sure looks like it, purge them all!
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!is.na(silly_code))
```

There are now a total of **11,214** GOD wells, as of 2025-04-02.

## Investigate GOD Samples With Existing Otolith Reads

No GOD samples should have existing otolith reads.

Of these 194 samples, how are they spread across DWPs?
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!is.na(otolith_mark_present)) %>% 
  dplyr::count(silly_code, `48_DWP_tray_barcode`) %>% 
  dplyr::arrange(dplyr::desc(n))
```

57 unique DWPs!!!

*see how egregious these are and if we need to do something about it*

### 22629

Start with the worst case. Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR17. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 22629)
```

Huh...there are only otolith transfers for 5 vial otoliths (fell out of DWP or stuck between acetate and DWP), but **not** for any DWP otoliths. That is weird...

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000022629") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

This DWP has complete otolith reads for nearly all 48 samples for PGILMOUR17 from 2017-09-06. The pattern of missing otoliths does not perfectly match "DWP maps", so I do not think that this DWP accidentally got put in the "unaffected" pile.

Check contractor data.
```{r}
stream_specimens %>% 
  dplyr::filter(sample_tray_id %in% c("0000022629", "0000022639")) %>% 
  dplyr::distinct(stream_name, survey_date, sample_tray_id)
```

Also has this otolith tray listed as PGILMOUR17, same date.

I suspect that the DWP barcode was incorrectly entered when otolith were initially read.

**See e-mail exchange with Tim Frawley 2022-04-26**

I'd found two DWPs coded as 0000022639, both PSTOCK17.
    
    Sample ID: 201799955 , (otolith) Tray Number: 0000022629, current DNA Tray Code: 0000022639 should be changed to DNA Tray Code: 0000022629

Otoliths from Sample ID 201799955 were read by Josh Buza in CDV on 2021-12-07, which checks out based on when this PGILMOUR17 DWP would have been shipped from MTAL to CDV after affected otoliths were removed for GOD.

However, the otolith reads associated with these samples in `AHRP_oceanAK` have Sample ID 53932_2900-53932_2947...

Any DWPs missing all otolith reads?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(is.na(otolith_mark_present)) %>% 
  dplyr::count(dna_tray_code, silly_code, sample_date) %>% 
  dplyr::filter(n == 48) %>% 
  dplyr::arrange(dna_tray_code)
```

None appear to be obvious.

#### Conclusion

2020-02 this DWP was shipped to MTAL and part of GOD
2021-03-04 this DWP was marked for GOD in `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`
2021-04 to 2021-10 no otoliths from wells in this DWP were transferred to otolith trays, not found in `data/otolith_transfer_final.csv`, however, 5 vial samples (not associated with a well) from this DWP were transferred to otolith trays
2021-12-07 `TRAY_NUMBER` 0000022629 was read by CDV (date matches GOD timeline for affected DWPs), but with `DNA_TRAY_CODE` 0000022639
2022-04-26 I had Tim Frawley change `DNA_TRAY_CODE` to 0000022629 for `SAMPLE_ID` 201799955

I do not have any record of GOD otoliths being transferred from the GOD DWP wells to 96-well otolith trays. Perhaps it accidently got put in the "unnafected" pile, but the pattern of missing otoliths in the read data does not match "DWP maps". Perhaps the "DWP maps" are off by a bit and some of the 0 wells actually had something in there? However, there were 5 vial otoliths from this DWP that did get transferred, 3 of which were assigned to hearts from PGILMOUR17_2900-2947.

Tried talking to Rachel Ertz (CDV), but unable to reconcile. Ultimately, I think we'll have to drop those 3 GOD vial matches **and** wipe out otolith data for 0000022629 from final pedigree analysis.

```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_oceanAK %>%
    dplyr::filter(dna_tray_code == "0000022629") %>%
    dplyr::select(dna_tray_code, dna_tray_well_code)
)
```

### 22749

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR17. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 22749)
```

This is another instance of a DWP that only had vial otoliths transferred (similar to 22629)! There are no DWP + well otoliths in here?

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000022749") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

I think this DWP got sorted into the *unaffected* pile and was read normally. Cross referencing with "DWP maps", it is missing data for the wells that had 0 otoliths, but otherwise has reads for most other wells.

Check contractor data.
```{r}
stream_specimens %>% 
  dplyr::filter(sample_tray_id %in% c("0000022749")) %>% 
  dplyr::arrange(sample_cell)
```

None of these wells were missing 2 otoliths in the field.

#### Conclusion

This DWP was accidently put in the "unaffected" by GOD pile and read normally. Need to wipe data from GOD wells (red cells), keep existing reads otherwise.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 22749) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

### 2202

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 2202)
```

There is no otolith transfer data, I suspect that these were accidentally put in the "unaffected" by GOD pile and read as is.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000002202") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

I think this DWP got sorted into the *unaffected* pile and was read normally. Cross referencing with "DWP maps", it is missing data for the well that had 0 otoliths (21) and wells with >2 otoliths (19, 27, 36, 37), but otherwise has reads for most other wells.

#### Conclusion

This DWP was accidentally put in the "unaffected" by GOD pile and read normally. Need to wipe data from GOD wells.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 2202) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

### 22679

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR17. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 22679)
```

Uhhh, these are **not** the wells that were supposed to be transferred based on "DWP maps"?

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000022679") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

#### Conclusion

There are 16 missing otoliths. The 14 listed above in the transfers + A4 (0 in "DWP maps") and B4 (1 in "DWP maps"). I think we can just wipe the original, marked, GOD otoliths and they try to re-pair the otoliths that were actually transferred by mistake? I'll need to add the actual transfers to `GOD_48DWP_wells_oceanak` so we can re-pair them.

Delete otolith read data from original GOD marked wells.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 22679) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

Remove original 14 GOD marked wells, replace with 14 mistakenly transferred wells.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(`48_DWP_tray_barcode` != 22679) %>%  # remove original, red marked GOD wells
  dplyr::bind_rows(PGOD21_48DWP_paired_genotype_matches %>% 
                     dplyr::filter(`48_DWP_tray_barcode` == 22679) %>%  # actual wells transferred for GOD
                     dplyr::mutate(`48_DWP_tray_barcode` = as.character(`48_DWP_tray_barcode`)) %>% 
                     dplyr::select(`48_DWP_tray_barcode`, `48_DWP_tray_position`) %>% 
                     dplyr::rename(heart_position = `48_DWP_tray_position`) %>% 
                     dplyr::mutate(
                       dna_tray_code = stringr::str_pad(string = `48_DWP_tray_barcode`, side = "left", width = 10, pad = "0")
                     ) %>%  # need 10 digit barcode, including leading 0's
                     dplyr::left_join(y = pos_to_code, by = dplyr::join_by(heart_position == pos)) %>%
                     dplyr::left_join(
                       y = AHRP_oceanAK,
                       by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
                     )  # join in Salmon Biological Fact Data, code from above
                   )
```

### 14596

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 14596)
```

Yes, otolith transfers occurred as expected.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000014596") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

I was expecting reads for wells 1-10 and 41-48, but not anything for 11-40 since those were marked for GOD. In particular, I'm skeptical of 23 and 24, since 24 was supposed to be empty (0 otoliths), same goes for 32.

#### Conclusion

Delete otolith read data from original GOD marked wells.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 14596) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

### 2839

See above under "Verify Otolith Transfers In GOD Wells"

    Only 1 vial otolith transferred, no red cells in "DWP maps", discordant GOD assignment and existing otolith read. Decided to mark red cells and will ask for otolith data to be deleted.

#### Conclusion

Delete otolith read data from 10 GOD marked wells.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 2839) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

### 2197

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 2197)
```

Uh oh, transfers did not occur as expected. This DWP was marked for GOD for columns 2-5, but it looks like 3-6 got transferred.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000002197") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

Okay, existing data matches the otolith transfer data, there are otolith reads for DWP columns 1-2, except for well 13, which was marked as 0 otoliths in "DWP maps".

#### Conclusion

I think we can just wipe the original, marked, GOD otoliths and they try to re-pair the otoliths that were actually transferred by mistake? I'll need to add the actual transfers to `GOD_48DWP_wells_oceanak` so we can re-pair them.

Delete otolith read data from original GOD marked wells.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` == 2197,
                      dna_tray_well_code %in% 9:16) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

Remove original GOD marked wells (columns 2-5), replace with mistakenly transferred wells (columns 3-6).
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(`48_DWP_tray_barcode` != 2197) %>%  # remove original, red marked GOD wells
  dplyr::bind_rows(dplyr::tibble(`48_DWP_tray_barcode` = "2197",
                                 dna_tray_well_code = 17:48) %>% # actual wells transferred for GOD
                     dplyr::left_join(y = pos_to_code, by = dplyr::join_by(dna_tray_well_code)) %>% 
                     dplyr::rename(heart_position = pos) %>%
                     dplyr::mutate(
                       dna_tray_code = stringr::str_pad(string = `48_DWP_tray_barcode`, side = "left", width = 10, pad = "0")
                     ) %>%  # need 10 digit barcode, including leading 0's
                     dplyr::left_join(
                       y = AHRP_oceanAK,
                       by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
                     )  # join in Salmon Biological Fact Data, code from above
                   )
```

### 16107

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 16107)
```

This DWP lost a lot of otoliths, hence all the vials. It looks like samples were removed from columns 2-5, but not column 6.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000016107") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

Yup, indeed that is what happened.

Check contractor data for missing otoliths.
```{r}
stream_specimens %>% 
  dplyr::filter(sample_tray_id %in% c("0000016107")) %>% 
  dplyr::arrange(sample_cell)
```

It looks like we were expecting H6 to be missing 1 otolith (matches "DWP maps"), but A6 should have had 0 otoliths and "DWP maps" shows 2. However, the existing otolith read data shows that A6 (41) was missing both otoliths, matching the contractor data.

#### Conclusion

Column 6 was supposed to be in GOD, but never got transferred. I'm okay with the existing otolith read data since all are the same read (wild) and it matches the contractor data.

Remove column 6 from GOD wells.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!(`48_DWP_tray_barcode` == 16107 & dna_tray_well_code %in% 41:48))  # remove column 6 from GOD
```

### 14415

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 14415)
```

Yes, otolith transfers occurred as expected, except G3 and H3 (23:24), which both had 1 otolith.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000014415") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

It looks like there are otolith reads for wells 23:24 and 30:32, which were GOD wells. All existing reads are wild.

Check contractor data for missing otoliths.
```{r}
stream_specimens %>% 
  dplyr::filter(sample_tray_id %in% c("0000014415")) %>% 
  dplyr::arrange(sample_cell)
```

Hrmm, nothing there.

Manually looked at the GOD otolith reads, only 1 was hatchery-origin, F3, 22. Unfortunately, this was not paired in GOD.

#### Conclusion

Remove wells 23:24 and 30:32 from GOD, keep existing otolith read data.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!(`48_DWP_tray_barcode` == 14415 & dna_tray_well_code %in% c(23:24, 30:32)))
```

### 14402

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 14402)
```

All otolith transfers occurred as expected.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000014402") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

#### Conclusion

This is a very clean DWP. Remove wells 13:15, 21 from GOD, keep existing otolith read data.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>%
  dplyr::filter(!(`48_DWP_tray_barcode` == 14402 & dna_tray_well_code %in% c(13:15, 21)))
```

### 16100

Confirmed that this DWP was supposedly in GOD (in "oto unpacking data", "DWP maps", and "pre-shipment inventory" worksheets of `data/raw_working_oto_sheets/working_oto_sheet_combined.xlsm`) as PGILMOUR15. We used barcode scanners, so I feel pretty confident about them.

Confirm otolith transfers occurred.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(`48_DWP_tray_barcode` == 16100)
```

Looks like row F was not transferred.

What is the existing data look like?
```{r}
AHRP_oceanAK %>% 
  dplyr::filter(dna_tray_code == "0000016100") %>% 
  dplyr::select(heart_fish_id, sample_date, dna_tray_code, dna_tray_well_code, sample_id:otolith_mark_status_code) %>% 
  dplyr::arrange(dna_tray_well_code)
```

Wells 14, 22, 30, and 38 were GOD wells, but were not transferred and got read. All were wild, there are some hatchery otoliths nearby, but given that none of the assigned GOD otoliths moved, I'm okay keeping these.

#### Conclusion

This is a very clean DWP. Remove wells 14, 22, 30, and 38 from GOD, keep existing otolith read data.
```{r}
GOD_48DWP_wells_oceanak <- GOD_48DWP_wells_oceanak %>% 
  dplyr::filter(!(`48_DWP_tray_barcode` == 16100 & dna_tray_well_code %in% c(14, 22, 30, 38)))
```

### All Remaining DWPs

All remaining 46 DWPs have 3 or fewer GOD wells with otolith data and it is not worth the time to go through them individually. I'm making the executive decision to delete the existing otolith read data for these wells. This is the most conservative decision, better to have good data than more data.
```{r}
# Get 46 GOD DWPs with 3 or less wells that have existing otolith data
GOD_48DWPs_3orless <- GOD_48DWP_wells_oceanak %>% 
  dplyr::anti_join(y = AHRP_wells_to_delete_otolith_data, by = dplyr::join_by(dna_tray_code, dna_tray_well_code)) %>%  # remove wells to be deleted from SALMON_BIO_FACT
  dplyr::filter(!is.na(otolith_mark_present)) %>% 
  dplyr::count(`48_DWP_tray_barcode`) %>% 
  dplyr::filter(n <= 3) %>% 
  dplyr::pull(`48_DWP_tray_barcode`)

# Add these 57 DWP wells to the 172 to be deleted
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::bind_rows(
      GOD_48DWP_wells_oceanak %>%
        dplyr::filter(`48_DWP_tray_barcode` %in% GOD_48DWPs_3orless,
                      !is.na(otolith_mark_present)) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::arrange(dna_tray_code, dna_tray_well_code)
)
```

There are 229 samples (DWP wells) that need to have otolith data deleted.

## GOD Wells to Delete Otolith Data

Add extra columns to help Eric L., CDV, and MTAL identify these samples in their respective databases.
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::distinct() %>% 
    dplyr::left_join(
      y = AHRP_oceanAK %>% dplyr::select(
        dna_tray_code,
        dna_tray_well_code,
        sample_id,
        silly_code,
        fish_id
      ),
      by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::rename(GCL_silly_code = silly_code, GCL_fish_id = fish_id)
)
```

How many unique DWPs?
```{r}
dplyr::n_distinct(AHRP_wells_to_delete_otolith_data$dna_tray_code)
```

**229 wells across 53 different DWPs with existing otolith data to delete**

# Verify No Otolith Reads for GOD Wells

Confirm that we've dealt with all GOD wells that had otolith read data.
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::anti_join(y = AHRP_wells_to_delete_otolith_data, by = dplyr::join_by(dna_tray_code, dna_tray_well_code)) %>%  # remove wells to be deleted from SALMON_BIO_FACT
  dplyr::filter(!is.na(otolith_mark_present))  # any remaining GOD wells left with otolith reads?
```

Excellent, we did it! Once Eric Lardizabal is able to delete the otolith data from those DWP wells, we'll be able to upload the PGOD21 matches.

# Verify PGOD21 Assignments Are GOD Wells

Now that we've trimmed which wells are in GOD and or not, let's return to verifying whether we have assigned any PGOD21 otoliths to non-GOD DWP wells. This will be a good double check to make sure we have correctly identified **ALL** the GOD wells. Some wells that were not marked for GOD, may have accidentally been transferred, becoming part of GOD.

The alternative is that our otolith transfer rules left some to be desired and perhaps we had some otoliths in PGOD21 that assigned to non-GOD wells that were deemed trustworthy.
```{r}
# wells
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position))

# count of wells by DWP
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
  dplyr::count(`48_DWP_tray_barcode`) %>% 
  dplyr::arrange(dplyr::desc(n))
```

There are still 55 distinct DWPs that have a total of 68 PGOD21 genotype assignments to non-GOD wells.

Did those otoliths move? In other words, were these unmarked GOD wells from which otoliths were mistakenly transferred, but matched themselves?
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>%  # matches not in GOD wells
  dplyr::count(otolith_moved)
```

## Add Unmoved PGOD21 Assignments To GOD Wells

It looks like otoliths from some unmarked wells accidentally got transferred for PGOD21. At this point, we only care if those PGOD21 otoliths were assigned to a heart genotype. If so, let's add the 41 unmoved otoliths (i.e. `otolith_moved = FALSE`) to `GOD_48DWP_wells_oceanak`. We'll filter out the remaining 23 moved + 4 NA otoliths in `07_verify-assigned-hearts-no-existing-otolith-read.Rmd`.

First, check and see if any have existing otolith reads.
```{r}
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>%  # matches not in GOD wells
  dplyr::filter(otolith_moved == FALSE) %>%  # only otoliths that did not move
  dplyr::select(`48_DWP_tray_barcode`, heart_position, oto_tray_and_well) %>%
    dplyr::mutate(
      dna_tray_code = stringr::str_pad(string = `48_DWP_tray_barcode`, side = "left", width = 10, pad = "0")
    ) %>%  # need 10 digit barcode, including leading 0's
    dplyr::left_join(y = pos_to_code, by = dplyr::join_by(heart_position == pos)) %>%
    dplyr::left_join(
      y = AHRP_oceanAK,
      by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
    ) %>% 
  dplyr::filter(!is.na(otolith_mark_present))
```

There are 4 of these 41 unmoved otoliths that have existing otolith reads. I manually checked with the GOD reads and they are the same!

Add those 4 unmoved otoliths with existing otolith reads to `AHRP_wells_to_delete_otolith_data`
```{r}
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::select(dna_tray_code, dna_tray_well_code) %>%
    dplyr::bind_rows(
      PGOD21_48DWP_paired_genotype_matches %>%
        dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
        dplyr::anti_join(
          y = GOD_48DWP_wells_oceanak %>%
            dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
          by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)
        ) %>%  # matches not in GOD wells
        dplyr::filter(otolith_moved == FALSE) %>%  # only otoliths that did not move
        dplyr::select(`48_DWP_tray_barcode`, heart_position, oto_tray_and_well) %>%
        dplyr::mutate(
          dna_tray_code = stringr::str_pad(
            string = `48_DWP_tray_barcode`,
            side = "left",
            width = 10,
            pad = "0"
          )
        ) %>%  # need 10 digit barcode, including leading 0's
        dplyr::left_join(y = pos_to_code, by = dplyr::join_by(heart_position == pos)) %>%
        dplyr::left_join(
          y = AHRP_oceanAK,
          by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
        ) %>%
        dplyr::filter(!is.na(otolith_mark_present)) %>%
        dplyr::select(dna_tray_code, dna_tray_well_code)
    ) %>%  # add 4 to delete
    dplyr::distinct() %>%
    dplyr::left_join(
      y = AHRP_oceanAK %>% dplyr::select(
        dna_tray_code,
        dna_tray_well_code,
        sample_id,
        silly_code,
        fish_id
      ),
      by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
    ) %>%
    dplyr::rename(GCL_silly_code = silly_code, GCL_fish_id = fish_id)
)
```

Add to `GOD_48DWP_wells_oceanak`.
```{r}
GOD_48DWP_wells_oceanak <-
  GOD_48DWP_wells_oceanak %>%
  dplyr::bind_rows(
    PGOD21_48DWP_paired_genotype_matches %>%
      dplyr::mutate(`48_DWP_tray_barcode` = as.character(`48_DWP_tray_barcode`)) %>%
      dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
      dplyr::anti_join(
        y = GOD_48DWP_wells_oceanak,
        by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)
      ) %>%  # matches not in GOD wells
      dplyr::filter(otolith_moved == FALSE) %>%  # only otoliths that did not move (i.e., self assignments)
      dplyr::select(`48_DWP_tray_barcode`, heart_position) %>%
      dplyr::mutate(
        dna_tray_code = stringr::str_pad(
          string = `48_DWP_tray_barcode`,
          side = "left",
          width = 10,
          pad = "0"
        )
      ) %>%  # need 10 digit barcode, including leading 0's
      dplyr::left_join(y = pos_to_code, by = dplyr::join_by(heart_position == pos)) %>%
      dplyr::left_join(
        y = AHRP_oceanAK,
        by = dplyr::join_by(dna_tray_code, dna_tray_well_code)
      )  # join in Salmon Biological Fact Data, code from above
  )
```

# Final QA/QC

## GOD Wells

### No Otolith Reads

Confirm that we've dealt with all GOD wells that had otolith read data, once we delete untrustworthy samples.
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::anti_join(y = AHRP_wells_to_delete_otolith_data, by = dplyr::join_by(dna_tray_code, dna_tray_well_code)) %>%  # remove wells to be deleted from SALMON_BIO_FACT
  dplyr::filter(!is.na(otolith_mark_present))  # any remaining GOD wells left with otolith reads?
```

Excellent, we did it! Once Eric Lardizabal is able to delete the otolith data from those DWP wells, we'll be able to upload the PGOD21 matches.

### All Assignments to GOD Wells

All PGOD21 assignments in GOD wells (except remaining 23 moved + 4 NA otoliths)
```{r}
# wells
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position))

# count of wells by DWP
PGOD21_48DWP_paired_genotype_matches %>% 
  dplyr::filter(!is.na(heart_position)) %>%  # filter for assignments
  dplyr::anti_join(y = GOD_48DWP_wells_oceanak %>% 
                     dplyr::mutate(`48_DWP_tray_barcode` = as.numeric(`48_DWP_tray_barcode`)),
                   by = dplyr::join_by(`48_DWP_tray_barcode`, heart_position)) %>% 
  dplyr::count(otolith_moved)
```

Great, that is the same 27 we expect. We'll deal with those in `07_verify-assigned-hearts-no-existing-otolith-read.Rmd`.

### No Duplicates

Remove any duplicated rows?
```{r}
nrow(GOD_48DWP_wells_oceanak) == GOD_48DWP_wells_oceanak %>% dplyr::distinct(`48_DWP_tray_barcode`, heart_position) %>% nrow()
```

## GOD Wells to Delete Otolith Data

Verify all have existing otolith reads and no duplicates.
```{r}
# what are reads for wells to delete
AHRP_wells_to_delete_otolith_data %>% 
  dplyr::distinct(dna_tray_code, dna_tray_well_code) %>%  # no duplicates
  dplyr::left_join(y = AHRP_oceanAK, by = dplyr::join_by(dna_tray_code, dna_tray_well_code)) %>%  # join in existing otolith data
  dplyr::count(otolith_mark_present)  # all have reads?
```

Yikes, 42 of our 233 wells do not have existing otolith read data. Dig into these 42 quickly.
```{r}
# filter for 42 wells
AHRP_wells_to_delete_otolith_data %>% 
  dplyr::distinct(dna_tray_code, dna_tray_well_code) %>%  # no duplicates
  dplyr::left_join(y = AHRP_oceanAK, by = dplyr::join_by(dna_tray_code, dna_tray_well_code)) %>%  # join in existing otolith data
  dplyr::filter(is.na(otolith_mark_present))
```

Okay, after spot checking a few of these above, I just included all GOD wells and did not filter for only those with reads. Correct here and then write out.
```{r}
# remove wells without existing otolith reads
(
  AHRP_wells_to_delete_otolith_data <- AHRP_wells_to_delete_otolith_data %>%
    dplyr::distinct() %>%  # no duplicates
    dplyr::left_join(
      y = AHRP_oceanAK,
      by = dplyr::join_by(dna_tray_code, dna_tray_well_code, sample_id)
    ) %>%  # join in existing otolith data
    dplyr::filter(!is.na(otolith_mark_present)) %>%
    dplyr::select(
      dna_tray_code,
      dna_tray_well_code,
      sample_id,
      GCL_silly_code,
      GCL_fish_id
    )
)
```

How many per collection?
```{r}
AHRP_wells_to_delete_otolith_data %>% 
  dplyr::count(GCL_silly_code)
```

How many DWPs?
```{r}
AHRP_wells_to_delete_otolith_data %>% 
  dplyr::count(dna_tray_code)
```

55 DWPs with data to delete.

# Write

How many GOD samples per collection?
```{r}
GOD_48DWP_wells_oceanak %>% 
  dplyr::count(silly_code)
```

Write out.
```{r}
# GOD wells
GOD_48DWP_wells_oceanak %>% 
  dplyr::select(`48_DWP_tray_barcode`, heart_position) %>% 
  readr::write_csv(file = "../output/GOD_48DWP_wells.csv")

# Otolith reads to delete, untrustworthy
readr::write_csv(x = AHRP_wells_to_delete_otolith_data, file = "../output/AHRP_GOD_wells_to_delete_otolith_read_data.csv")
```

# Future Work

The next steps after this will be to send Eric L. database changes: `AHRP_wells_to_delete_otolith_data`. Once those have been made to the respective CDV and MTAL databases, it should update in Salmon Biological Fact. Then I'll pull the updated Salmon Biological Fact data from the warehouse and  *investigate GOD otoliths with reads but no heart match*, using `GOD_48DWP_wells.csv` to determine how many wells have unpaired heart/otolith data and see if we can assign remaining PGOD21 otolith reads to wells not genetically assigned to a heart. Following that we'll need to *verify heart matches are GOD fish (i.e., don't have existing otolith reads)* and send finalized matches to Eric L.

# Summary

There were originally 10,898 cells with red font in `data\raw_working_oto_sheets\working_oto_sheet_combined.xlsm`, which were ostensibly GOD DWP wells. I copied the "DWP maps" tab into a separate spreadsheet `data\raw_working_oto_sheets\DWP_maps_clean.xlsm` and did extensive QA/QC of the GOD wells, extracting a list of red font cells with a VBA macro written with ChatGPT. The QA/QC process involved: removing unsampled wells (partially filled DWPs, otoliths migrated to unsampled wells), removing any wells that had existing otolith data and were deemed to be trustworthy, cross referencing GOD wells with otolith transfer data to add missing DWPs and wells, cross referencing GOD wells with existing otolith reads in Salmon Biological Fact to update "DWP maps", adding the 7 DWPs from in-house GCL otolith transfers, and identifying 233 wells that have existing otolith data that need to be deleted due untrustworthy data (i.e., DWP accidentally got sorted into the unaffected pile and read as is, wrong otoliths were transferred, etc.). I went through individual DWPs that had >3 GOD wells with existing otolith data to resolve discrepancies, but decided to existing otolith data from 57 wells from 46 different DWPs that had existing otolith data for <=3 GOD wells. I also added 41 GOD wells for otoliths that accidentally got transferred and self-assigned (i.e., `otolith_moved == FALSE`).

This leaves **11,234** wells that are actually GOD wells. Once the **191** wells are deleted, no GOD wells will have existing otolith reads. Note that there are still 27 PGOD21 otoliths assigned to non-GOD wells, but we'll deal with this in `07_verify-assigned-hearts-no-existing-otolith-read`, and drop those assignments.

End
